Add run-dredd-command hook
Add run-dredd-command hook making it possible to execute actions
before dredd testing starts.
