+++ extensions/roc-plugin-dredd/src/roc/index.js
@@ -1,5 +1,5 @@
 import { lazyFunctionRequire } from 'roc';
-import { isBoolean } from 'roc/validators';
+import { isBoolean, required, notEmpty } from 'roc/validators';
 
 import config from '../config/roc.config.js';
 import meta from '../config/roc.config.meta.js';
@@ -9,6 +9,13 @@ const lazyRequire = lazyFunctionRequire(require);
 export default {
     config,
     meta,
+    actions: [
+        {
+            hook: 'run-dredd-command',
+            description: 'Run dredd command',
+            action: lazyRequire('../actions/dredd'),
+        },
+    ],
     commands: {
         development: {
             dredd: {
@@ -27,6 +34,15 @@ export default {
         },
     },
     hooks: {
+        'run-dredd-command': {
+            description: 'Used to run the dredd command',
+            arguments: {
+                watch: {
+                    validator: required(notEmpty(isBoolean)),
+                    description: 'Run dredd in watch mode',
+                },
+            },
+        },
         'run-dev-command': {
             description: 'Used to start dev server used for dredd testing',
         },
